Switch from `float` to `double` for improved precision across domain models, generators, and tests. Update associated logic, constructors, validations, and tests to reflect the change.

+++ src/main/java/galaxy/domain/Coordinates.java
@@ -3,17 +3,17 @@ package galaxy.domain;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.jme3.math.Vector3f;
 
-public record Coordinates(float x, float y, float z) {
+public record Coordinates(double x, double y, double z) {
 
 	public Coordinates(Vector3f source) {
 		this(source.x, source.y, source.z);
 	}
 
 	public Coordinates(JsonNode source) {
-		this(source.path("x").floatValue(), source.path("y").floatValue(), source.path("z").floatValue());
+		this(source.path("x").doubleValue(), source.path("y").doubleValue(), source.path("z").doubleValue());
 	}
 
 	public Vector3f asVector3f() {
-		return new Vector3f(x, y, z);
+		return new Vector3f((float) x, (float) y, (float) z);
 	}
 }
