+++ src/main/java/galaxy/domain/Industry.java
@@ -2,15 +2,15 @@ package galaxy.domain;
 
 import com.fasterxml.jackson.databind.JsonNode;
 
-public record Industry(float value) {
+public record Industry(double value) {
 
 	public Industry {
-		if (value < 0f) {
+		if (value < 0) {
 			throw new IllegalArgumentException("Industry value must be positive");
 		}
 	}
 
 	public Industry(JsonNode source) {
-		this(source.floatValue());
+		this(source.doubleValue());
 	}
 }
