+++ src/main/java/galaxy/domain/Population.java
@@ -2,15 +2,15 @@ package galaxy.domain;
 
 import com.fasterxml.jackson.databind.JsonNode;
 
-public record Population(float value) {
+public record Population(double value) {
 
 	public Population {
-		if (value < 0f) {
+		if (value < 0) {
 			throw new IllegalArgumentException("Population value must be positive");
 		}
 	}
 
 	public Population(JsonNode source) {
-		this(source.floatValue());
+		this(source.doubleValue());
 	}
 }
